+++ lib/rspec/activerecord/expectations.rb
@@ -8,6 +8,22 @@ module RSpec
       def repeatedly_load(klass)
         Matchers::LoadMatcher.new(klass)
       end
+
+      def execute_a_transaction
+        Matchers::TransactionMatcher.new(:transaction_queries)
+      end
+
+      def rollback_a_transaction
+        Matchers::TransactionMatcher.new(:rollback_queries)
+      end
+
+      def roll_back_a_transaction
+        Matchers::TransactionMatcher.new(:rollback_queries)
+      end
+
+      def commit_a_transaction
+        Matchers::TransactionMatcher.new(:commit_queries)
+      end
     end
   end
 end
@@ -17,3 +33,4 @@ require_relative 'expectations/query_inspector'
 require_relative 'expectations/collector'
 require_relative 'expectations/matchers/query_count_matcher'
 require_relative 'expectations/matchers/load_matcher'
+require_relative 'expectations/matchers/transaction_matcher'